+++ timecode.h
@@ -6,6 +6,6 @@
 long long int do_drop_frames (double fps, int f, int s, int m, int h);
 long long int bcd_to_framecnt(double fps, int df, int f, int s, int m, int h);
 long long int ltcframe_to_framecnt(LTCFrame *lf, double fps);
-long long frame_to_ms(LTCFrame *f, int fps_num, int fps_den);
+double frame_to_ms(LTCFrame *f, int fps_num, int fps_den);
 
 #endif
